Phase 4: extract radar layout math, legend, and blip dot from RadarChart
